+++ src/main/java/com/diaz/expense_tracker/service_impl/ExpenseServiceImpl.java
@@ -1,4 +1,4 @@
-package com.diaz.expense_tracker.impl;
+package com.diaz.expense_tracker.service_impl;
 
 import com.diaz.expense_tracker.Repository.ExpenseRepository;
 import com.diaz.expense_tracker.model.Expense;
