+++ src/main/java/com/diaz/expense_tracker/service_impl/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.diaz.expense_tracker.impl;
+package com.diaz.expense_tracker.service_impl;
 
 import com.diaz.expense_tracker.Repository.UserRepository;
 import com.diaz.expense_tracker.model.User;
